# docs/rebuild/openapi-baseline.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Tradier Bot API",
  "description": "Standalone backend for three worlds on one process: the Tradier options desk (balance, chain preview, managed positions with TP/SL, the opening-range and A/B auto-traders, the HOT scan and the options-flow board), the 36 Trade Desk that fronts it on a phone, and the Bot Station that launches, supervises and settles the vendored Kalshi bots (btc15, btc60, sports, parlay, gold15, silver15, oil15). Plus the super_research signal engines and the SPY/QQQ/SPX level watcher they read. SQLite persistence, per-user credential folders, paper-only by default.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/bots/btc/processes": {
   "get": {
    "tags": [
     "bots"
    ],
    "summary": "Btc Processes",
    "description": "Stray-process check: what is actually running on this machine.",
    "operationId": "getBtcBotProcesses",
    "parameters": [
     {
      "name": "bot",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "btc15 or btc60",
       "default": "btc15",
       "title": "Bot"
      },
      "description": "btc15 or btc60"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Getbtcbotprocesses"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/bots/btc/kill": {
   "post": {
    "tags": [
     "bots"
    ],
    "summary": "Btc Kill",
    "description": "Kill every process running this BTC bot \u2014 including copies the API did\nnot start. Use when a stray run blocks a start.",
    "operationId": "killBtcBotProcesses",
    "parameters": [
     {
      "name": "bot",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "btc15",
       "title": "Bot"
      }
     },
     {
      "name": "pids",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "comma-separated; omit to kill all",
       "title": "Pids"
      },
      "description": "comma-separated; omit to kill all"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Killbtcbotprocesses"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/bots/sports/kill": {
   "post": {
    "tags": [
     "bots"
    ],
    "summary": "Sports Kill",
    "operationId": "killSportsBotProcesses",
    "parameters": [
     {
      "name": "pids",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "comma-separated; omit to kill all",
       "title": "Pids"
      },
      "description": "comma-separated; omit to kill all"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Killsportsbotprocesses"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/bots/btc/status": {
   "get": {
    "tags": [
     "bots"
    ],
    "summary": "Btc Status",
    "operationId": "getBtcBotStatus",
    "parameters": [
     {
      "name": "user_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/BotStatusOut"
         },
         "title": "Response Getbtcbotstatus"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/bots/btc/start": {
   "post": {
    "tags": [
     "bots"
    ],
    "summary": "Btc Start",
    "operationId": "startBtcBot",
    "parameters": [
     {
      "name": "bot",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "btc15 or btc60",
       "default": "btc15",
       "title": "Bot"
      },
      "description": "btc15 or btc60"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BotStartRequest"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BotRunOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/bots/btc/stop": {
   "post": {
    "tags": [
     "bots"
    ],
    "summary": "Btc Stop",
    "operationId": "stopBtcBot",
    "parameters": [
     {
      "name": "bot",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "btc15 or btc60",
       "default": "btc15",
       "title": "Bot"
      },
      "description": "btc15 or btc60"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BotStopRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/BotRunOut"
         },
         "title": "Response Stopbtcbot"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/bots/btc/logs": {
   "get": {
    "tags": [
     "bots"
    ],
    "summary": "Btc Logs",
    "operationId": "getBtcBotLogs",
    "parameters": [
     {
      "name": "bot",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "btc15",
       "title": "Bot"
      }
     },
     {
      "name": "user_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "User Id"
      }
     },
     {
      "name": "run_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Run Id"
      }
     },
     {
      "name": "lines",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 2000,
       "minimum": 1,
       "default": 100,
       "title": "Lines"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BotLogsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/bots/btc/sync-trades": {
   "post": {
    "tags": [
     "bots"
    ],
    "summary": "Btc Sync Trades",
    "operationId": "syncBtcTrades",
    "parameters": [
     {
      "name": "user_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     },
     {
      "name": "bot",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Bot"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response Syncbtctrades"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/bots/sports/status": {
   "get": {
    "tags": [
     "bots"
    ],
    "summary": "Sports Status",
    "operationId": "getSportsBotStatus",
    "parameters": [
     {
      "name": "user_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BotStatusOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BotStatusOut": {
    "properties": {
     "bot_key": {
      "type": "string",
      "title": "Bot Key"
     },
     "running": {
      "type": "boolean",
      "title": "Running"
     },
     "runs": {
      "items": {
       "$ref": "#/components/schemas/BotRunOut"
      },
      "type": "array",
      "title": "Runs"
     },
     "session": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session"
     }
    },
    "type": "object",
    "required": [
     "bot_key",
     "running",
     "runs"
    ],
    "title": "BotStatusOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
  }
 }
}